API Documentation

API Documentation

This is the stable public API of BAT. Forward/backward compatibility follows Julia's semantic versioning rules.

Types

BAT.ARPWeighting

BAT.AbstractDensity

BAT.AbstractMCMCTuningStrategy

BAT.AbstractPosteriorDensity

BAT.AbstractSamplingAlgorithm

BAT.AbstractWeightingScheme

BAT.AdaptiveMetropolisTuning

BAT.BrooksGelmanConvergence

BAT.DensitySample

BAT.DensitySampleVector

BAT.DistLikeDensity

BAT.GelmanRubinConvergence

BAT.MCMCAlgorithm

BAT.MCMCBurninStrategy

BAT.MCMCInitStrategy

BAT.MCMCIterator

BAT.MetropolisHastings

BAT.NamedTupleDist

BAT.PosteriorDensity

BAT.RandResampling

BAT.RandSampling

BAT.RepetitionWeighting

Functions and macros

BAT.bat_integrate

BAT.bat_read

BAT.bat_rng

BAT.bat_sample

BAT.bat_stats

BAT.bat_write

BAT.default_sampling_algorithm

BAT.density_logval

BAT.nparams

Documentation

BAT.bat_integrate — Function.

bat_integrate(
    posterior::BAT.AnyPosterior,
)::DensitySampleVector

Calculate the integral (evidence) of posterior.

Returns a NamedTuple: (integral = x::Measurement.Measurement, ...)

Result properties not listed here are algorithm-specific and are not part of the stable BAT API.

posterior may be a

BAT.AbstractPosteriorDensity

BAT.DistLikeDensity

BAT.DensitySampleVector

Distributions.MultivariateDistribution

Uses the AHMI algorithm by default.

source

BAT.bat_read — Function.

bat_read(filename::AbstractString, data)
bat_read(fn_with_subpath::Tuple{AbstractString, AbstractString}, data)

Read data from a file filename, resp. from an internal sub-path of the file (if supported by the file format), e.g. an HDF5 group.

Currently supported file formats are:

HDF5 with file extension ".h5" or ".hdf5"

source

BAT.bat_rng — Function.

bat_rng()

Return a new BAT-compatigle random number generator, with a random seed drawn from the system entropy pool.

source

BAT.bat_sample — Function.

bat_sample(
    [rng::AbstractRNG],
    posterior::BAT.AnyPosterior,
    n::BAT.AnyNSamples,
    [algorithm::BAT.AbstractSamplingAlgorithm]
)::DensitySampleVector

Draw n samples from posterior.

Returns a NamedTuple of the shape

(
    samples = X::DensitySampleVector,...
    stats = s::@test stats isa NamedTuple{(:mode,:mean,:cov,...)},
    ...
)

Result properties not listed here are algorithm-specific and are not part of the stable BAT API.

posterior may be a

BAT.AbstractPosteriorDensity

BAT.DistLikeDensity

BAT.DensitySampleVector

Distributions.MultivariateDistribution

Depending on the type of posterior, n may be of type

Integer: Number of samples

Tuple{Integer,Integer}: Tuple of number of samples per sample source and number of sample sources (e.g. number of MCMC chains). The total number of samples is product(n).

Depending on the type of posterior, the number of samples returned may be somewhat larger or smaller than specified by product(n).

Also depending on the posterior type, the samples may be independent or correlated (e.g. when using MCMC).

source

BAT.bat_stats — Function.

bat_stats(samples::DensitySampleVector)

Calculated parameter statistics on samples. Returns a NamedTuple{(:mode,:mean,:cov,...)}. Result properties not listed here are not part of the stable BAT API and subject to change.

source

BAT.bat_write — Function.

bat_write(filename::AbstractString, data)
bat_write(fn_with_subpath::Tuple{AbstractString, AbstractString}, data)

Write data to a file filename, resp. to an internal sub-path of the file (if supported by the file format), e.g. an HDF5 group.

Currently supported file formats are:

HDF5 with file extension ".h5" or ".hdf5"

source

BAT.nparams — Function.

BAT.nparams(X::Union{AbstractParamBounds,MCMCIterator,...})

Get the number of parameters of X.

source

BAT.default_sampling_algorithm — Function.

BAT.default_sampling_algorithm(posterior)

Get BAT's default sampling algorithm for posterior.

source

BAT.density_logval — Function.

BAT.density_logval(density::AbstractDensity, params::Any)

Compute log of value of a multivariate density function at the given parameter values.

Input:

density: density function

params: parameter values

Note: If density_logval is called with out-of-bounds parameters (see param_bounds), the behaviour is undefined. The result for parameters that are not within bounds is implicitly -Inf, but it is the caller's responsibility to handle these cases.

source

BAT.AbstractDensity — Type.

AbstractDensity

Subtypes of AbstractDensity must implement the function

BAT.density_logval

For likelihood densities this is typically sufficient, since shape, and bounds of parameters will be inferred from the prior.

Densities with known parameters bounds may also implement

BAT.param_bounds

If the parameter bounds are unkown, but the number of parameters is known, the function

BAT.nparams

may be implemented directly (usually it is inferred from the bounds).

Note

The functions BAT.param_bounds and BAT.params_shape are currently not part of the stable public BAT-API, their names and arguments may change without notice.

Densities that support named parameters should also implement

BAT.params_shape

source

BAT.AbstractMCMCTuningStrategy — Type.

AbstractMCMCTuningStrategy

Abstract super-type for MCMC tuning strategies.

source

BAT.AbstractPosteriorDensity — Type.

abstract type AbstractPosteriorDensity <: AbstractDensity end

Abstract super-type for posterior probability densities.

source

BAT.AbstractWeightingScheme — Type.

AbstractWeightingScheme{T<:Real}

Abstract class for sample weighting schemes.

Weight values will have type T.

source

BAT.AdaptiveMetropolisTuning — Type.

AdaptiveMetropolisTuning(...)

Adaptive MCMC tuning strategy for Metropolis-Hastings samplers.

Adapts the proposal function based on the acceptance ratio and covariance of the previous samples.

Fields:

λ: Controls the weight given to new covariance information in adapting the proposal distribution. Defaults to 0.5.

α: Metropolis-Hastings acceptance ratio target, tuning will try to adapt the proposal distribution to bring the acceptance ratio inside this interval. Defaults to IntervalSets.ClosedInterval(0.15, 0.35)

β: Controls how much the spread of the proposal distribution is widened/narrowed depending on the current MH acceptance ratio.

c: Interval for allowed scale/spread of the proposal distribution. Defaults to ClosedInterval(1e-4, 1e2).

Constructors:

AdaptiveMetropolisTuning(
    λ::Real,
    α::IntervalSets.ClosedInterval{<:Real},
    β::Real,
    c::IntervalSets.ClosedInterval{<:Real}
)

source

BAT.ARPWeighting — Type.

struct ARPWeighting{T<:AbstractFloat} <: AbstractWeightingScheme{T}

Sample weighting scheme suitable for accept/reject-based sampling algorithms (e.g. MetropolisHastings). Both accepted and rejected samples become part of the output, with a weight proportional to their original acceptance probability.

Constructors:

ARPWeighting()

ARPWeighting{T}()

source

BAT.BrooksGelmanConvergence — Type.

BrooksGelmanConvergence

Brooks-Gelman $maximum(R^2)$ convergence test.

source

BAT.DensitySample — Type.

struct DensitySample

A weighted sample drawn according to an statistical density, e.g. a BAT.AbstractDensity.

Fields: * params: Multivariate parameter vector * logdensity: log of the value of the density at params * weight: Weight of the sample * info: Additional info on the provenance of the sample. Content depends on the sampling algorithm. * aux: Custom user-defined information attatched to the sample.

Constructors:

DensitySample(
    params::AbstractVector{<:Real},
    logdensity::Real,
    weight::Real,
    info::Any,
    aux::Any
)

Use DensitySampleVector to store vectors of multiple samples with an efficient column-based memory layout.

source

BAT.DensitySampleVector — Type.

DensitySampleVector

Type alias for StructArrays.StructArray{<:DensitySample,...}.

Constructor:

    DensitySampleVector(
        (
            params::AbstractVector{<:AbstractVector{<:Real}}
            logdensity::AbstractVector{<:Real}
            weight::AbstractVector{<:Real}
            info::AbstractVector{<:Any}
            aux::AbstractVector{<:Any}
        )
    )

source

BAT.DistLikeDensity — Type.

DistLikeDensity <: AbstractDensity

A density that implements part of the Distributions.Distribution interface. Such densities are suitable to be used as a priors.

Typically, custom priors should be implemented as subtypes of Distributions.Distribution. BAT will automatically wrap them in a subtype of DistLikeDensity.

Subtypes of DistLikeDensity are required to support more functionality than a AbstractDensity, but less than a Distribution{Multivariate,Continuous}.

The following functions must be implemented for subtypes:

BAT.density_logval

BAT.param_bounds

BAT.params_shape

Distributions.sampler

Statistics.cov

Note

The functions BAT.param_bounds and BAT.params_shape are currently not part of the stable public BAT-API, their names and arguments may change without notice.

Prior densities that support non-flat parameters should also implement

BAT.params_shape

A d::Distribution{Multivariate,Continuous} can be converted into (wrapped in) an DistLikeDensity via conv(DistLikeDensity, d).

source

BAT.GelmanRubinConvergence — Type.

GelmanRubinConvergence

Gelman-Rubin $\$maximum(R^2)\$$ convergence test.

source

BAT.MCMCAlgorithm — Type.

abstract type MCMCAlgorithm <: AbstractSamplingAlgorithm end

Note

The details of the MCMCIterator and MCMCAlgorithm API (see below) currently do not form part of the BAT-API, and may change without notice.

The following methods must be defined for subtypes (e.g. for SomeAlgorithm<:MCMCAlgorithm):

(spec::MCMCSpec{<:SomeAlgorithm})(chainid::Integer)::MCMCIterator

In some cases, these custom methods may be necessary (default methods are defined for MCMCAlgorithm):

BAT.initial_params!(
    params::Union{AbstractVector{<:Real},VectorOfSimilarVectors{<:Real}},
    rng::AbstractRNG,
    posterior::AbstractPosteriorDensity,
    algorithm::SomeAlgorithm
)

To implement a new MCMC algorithm, subtypes of both MCMCAlgorithm and MCMCIterator are required.

source

BAT.MCMCBurninStrategy — Type.

@with_kw struct MCMCBurninStrategy

Defines the MCMC burn-in strategy, specifically the number and length of MCMC tuning/burn-in cycles.

Fields:

max_nsamples_per_cycle: Maximum number of MCMC samples to generate per cycle, defaults to 1000. Definition of a sample depends on MCMC algorithm.

max_nsteps_per_cycle: Maximum number of MCMC steps per cycle, defaults to 10000. Definition of a step depends on MCMC algorithm.

max_time_per_cycle: Maximum wall-clock time to spend per cycle, in seconds. Defaults to Inf.

max_ncycles: Maximum number of cycles.

source

BAT.MCMCInitStrategy — Type.

@with_kw struct MCMCInitStrategy

Defines the MCMC chain initialization strategy.

Fields:

init_tries_per_chain: Interval that specifies the minimum and maximum number of tries per MCMC chain to find a suitable starting position. Many candidate chains will be created and run for a short time. The chains with the best performance will be selected for tuning/burn-in and MCMC sampling run. Defaults to IntervalSets.ClosedInterval(8, 128).

max_nsamples_init: Maximum number of MCMC samples for each candidate chain. Defaults to 25. Definition of a sample depends on sampling algorithm.

max_nsteps_init: Maximum number of MCMC steps for each candidate chain. Defaults to 250. Definition of a step depends on sampling algorithm.

max_time_init::Int: Maximum wall-clock time to spend per candidate chain, in seconds. Defaults to Inf.

source

BAT.MCMCIterator — Type.

abstract type MCMCIterator end

Represents the current state of a MCMC chain.

Note

The details of the MCMCIterator and MCMCAlgorithm API (see below) currently do not form part of the BAT-API, and may change without notice.

To implement a new MCMC algorithm, subtypes of both MCMCAlgorithm and MCMCIterator are required.

The following methods must be defined for subtypes of MCMCIterator (e.g. SomeMCMCIter<:MCMCIterator):

BAT.mcmc_spec(chain::SomeMCMCIter)::MCMCSpec

BAT.getrng(chain::SomeMCMCIter)::AbstractRNG

BAT.mcmc_info(chain::SomeMCMCIter)::MCMCIteratorInfo

BAT.nsteps(chain::SomeMCMCIter)::Int

BAT.nsamples(chain::SomeMCMCIter)::Int

BAT.current_sample(chain::SomeMCMCIter)::DensitySample

BAT.sample_type(chain::SomeMCMCIter)::Type{<:DensitySample}

BAT.samples_available(chain::SomeMCMCIter, nonzero_weights::Bool = false)::Bool

BAT.get_samples!(appendable, chain::SomeMCMCIter, nonzero_weights::Bool)::typeof(appendable)

BAT.next_cycle!(chain::SomeMCMCIter)::SomeMCMCIter

BAT.mcmc_step!(
    callback::AbstractMCMCCallback,
    chain::SomeMCMCIter
)

The following methods are implemented by default:

algorithm(chain::MCMCIterator)
getposterior(chain::MCMCIterator)
rngseed(chain::MCMCIterator)
nparams(chain::MCMCIterator)
DensitySampleVector(chain::MCMCIterator)
mcmc_iterate!(callback, chain::MCMCIterator, ...)
mcmc_iterate!(callbacks, chains::AbstractVector{<:MCMCIterator}, ...)

source

BAT.MetropolisHastings — Type.

MetropolisHastings

Metropolis-Hastings MCMC sampling algorithm. !!!!!

source

BAT.NamedTupleDist — Type.

NamedTupleDist <: MultivariateDistribution
NamedTupleDist <: MultivariateDistribution

A distribution with NamedTuple-typed variates.

Can be used to describe the distribution of each parameter in a set of named parameters. If the distribution is used as a Bayesian prior, the NamedTupleDist then specifies the prior on each named parameter.

A NamedTupleDist implies a NamedTupleShape:

valshape(d::NamedTupleDist) isa NamedTupleShape

source

BAT.PosteriorDensity — Type.

PosteriorDensity{
    Li<:AbstractDensity,
    Pr<:DistLikeDensity
} <: AbstractPosteriorDensity

A representation of a PosteriorDensity, based a likelihood and prior, a representation of the posterior density is cached internally. The densities be accessed via

getlikelihood(posterior::PosteriorDensity)::Li
getprior(posterior::PosteriorDensity)::Pr

Constructors:

PosteriorDensity(likelihood::AbstractDensity, prior::AbstractDensity)
PosteriorDensity(likelihood::Any, prior::Any)
PosteriorDensity(log_likelihood::Function, prior::Any)

source

BAT.RandResampling — Type.

RandResampling <: AbstractSamplingAlgorithm

Constructors:

RandResampling()

Resamples from a given set of samples.

source

BAT.RandSampling — Type.

RandSampling

Constructors:

RandSampling()

Sample via Random.rand. Only supported for posteriors of type Distributions.MultivariateDistribution and BAT.DistLikeDensity.

source

BAT.RepetitionWeighting — Type.

struct RepetitionWeighting{T<:AbstractFloat} <: AbstractWeightingScheme{T}

Sample weighting scheme suitable for sampling algorithms which may repeated samples multiple times in direct succession (e.g. MetropolisHastings). The repeated sample is stored only once, with a weight equal to the number of times it has been repeated (e.g. because a Markov chain has not moved during a sampling step).

Constructors:

RepetitionWeighting()

RepetitionWeighting{T}()

source

BAT.AbstractSamplingAlgorithm — Type.

BAT.AbstractSamplingAlgorithm

Abstract type for BAT sampling algorithms. See bat_sample.

source